requirements.txt
Django==5.2
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
drf-spectacular